services/telemetry_ingest/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
pyjwt==2.9.0
httpx==0.27.2
